Revert "Kan 37"
