+++ app.py
@@ -6,7 +6,7 @@ from odds_api import get_odds_data, get_sports
 from historical_odds import get_sdql_data
 from single_game_data import get_game_details
 from sdql_queries import get_last_5_games, get_last_5_games_vs_opponent
-from utils import convert_sport_key, mlb_totals, other_totals, convert_to_eastern, check_for_trends, convert_team_name
+from utils import convert_sport_key, mlb_totals, other_totals, convert_to_eastern, check_for_trends
 from betting_guide import betting_guide
 from flask_caching import Cache
 import logging
@@ -17,19 +17,8 @@ from celery_config import celery
 from tasks import show_trends_task  # Import the task from tasks module
 import subprocess
 import redis
-import requests
 from urllib.parse import urlparse, parse_qs
 import ssl
-import json
-#from shared_utils import convert_roto_team_names
-#from mlb_pitchers import get_starting_pitchers
-from scores_templates import (
-    historical_template,
-    default_template,
-    baseball_template,
-    soccer_template
-)
-from game_details_templates import mlb_template, nhl_template, others_template
 
 load_dotenv()  # Load environment variables from .env file
 
@@ -246,6 +235,74 @@ def task_status(task_id):
         }
     return jsonify(response)
 
+# @celery.task(name='app.show_trends_task')
+# def show_trends_task(sport_key, date):
+#     # Example task implementation
+#     selected_date_start = datetime.strptime(date, '%Y-%m-%d').replace(tzinfo=eastern_tz)
+#     scores, odds = get_odds_data(sport_key, selected_date_start)
+#     if scores is None or odds is None:
+#         return {'error': 'Error fetching odds data'}
+
+#     filtered_scores = []
+#     for score in scores:
+#         commence_time_str = score['commence_time']
+#         commence_date = parser.parse(commence_time_str).astimezone(pytz.utc)
+#         commence_date_eastern = convert_to_eastern(commence_date)
+
+#         if commence_date_eastern.date() == selected_date_start.date():
+#             filtered_scores.append(score)
+
+#     formatted_scores = []
+#     for match in filtered_scores:
+#         home_team = match.get('home_team', 'N/A')
+#         away_team = match.get('away_team', 'N/A')
+#         home_score = match['scores' ][0]['score'] if match.get('scores') else 'N/A'
+#         away_score = match['scores'][1]['score'] if match.get('scores') else 'N/A'
+
+#         match_odds = next((odds_match for odds_match in odds if odds_match['id'] == match['id']), None)
+#         odds_data = {'h2h': [], 'spreads': [], 'totals': []}
+
+#         if match_odds:
+#             for bookmaker in match_odds['bookmakers']:
+#                 for market in bookmaker['markets']:
+#                     market_key = market['key']
+#                     for outcome in market['outcomes']:
+#                         outcome_text = f"{outcome['name']}"
+#                         if market_key in ['spreads', 'totals'] and 'point' in outcome:
+#                             outcome_text += f": {outcome['point']}"
+#                         if market_key == 'h2h':
+#                             price = outcome['price']
+#                             if price > 0:
+#                                 outcome_text += f": +{price}"
+#                             else:
+#                                 outcome_text += f": {price}"
+#                         else:
+#                             price = outcome['price']
+#                             if price > 0:
+#                                 outcome_text += f" +{price}"
+#                             else:
+#                                 outcome_text += f" {price}"
+#                         odds_data[market_key].append(outcome_text)
+
+#         formatted_scores.append({
+#             'homeTeam': home_team,
+#             'awayTeam': away_team,
+#             'homeScore': home_score,
+#             'awayScore': away_score,
+#             'odds': odds_data,
+#             'game_id': match['id'],
+#         })
+    
+
+#     # Filter the games to include only those with trends
+#     games_with_trends = [game for game in formatted_scores if check_for_trends(game, selected_date_start, sport_key)['trend_detected']]
+
+#     return {
+#         'result': games_with_trends,
+#         'sport_key': sport_key,
+#         'current_date': date
+#     }
+
 # Route to fetch available sports
 @app.route('/api/sports')
 @cache.cached(timeout=3600, query_string=True)
@@ -295,137 +352,411 @@ def get_sport_scores(sport_key):
             current_date = datetime.now(eastern_tz).strftime('%Y-%m-%d')
 
         selected_date_start = datetime.strptime(current_date, '%Y-%m-%d').replace(tzinfo=eastern_tz)
+        selected_date_end = selected_date_start + timedelta(hours=23, minutes=59, seconds=59)
 
         sdql_sport_key = convert_sport_key(sport_key)
 
-        # Historical path
         if selected_date_start.date() < datetime.now(eastern_tz).date():
             sdql_data = get_sdql_data(sdql_sport_key, selected_date_start)
-            return render_template_string(historical_template, result=sdql_data)
+            return render_template_string("""
+                <html>
+                <head>
+                    <!-- Google tag (gtag.js) -->
+                    <script async src="https://www.googletagmanager.com/gtag/js?id=G-578SDWQPSK"></script>
+                    <script>
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
 
-        # Live/future path
-        scores, odds = get_odds_data(sport_key, selected_date_start)
-        if scores is None or odds is None:
-            return jsonify({'error': 'Error fetching odds data'}), 500
+                    gtag('config', 'G-578SDWQPSK');
+                    </script> 
+                    <script 
+                        async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6546677374101814"
+                        crossorigin="anonymous">
+                    </script>                     
+                    <title>Game Info</title>
+                    <style>
+                        table {
+                            width: 50%;
+                            border-collapse: collapse;
+                        }
+                        table, th, td {
+                            border: 1.5px solid #ddd;
+                        }
+                        th, td {
+                            padding: 8px;
+                            text-align: left;
+                        }
+                        th {
+                            background-color: #f2f2f2;
+                        }
+                        .game-pair {
+                            margin-bottom: 20px;
+                        }
+                    </style>
+                </head>
+                <body>
+                    <h1>Game Information</h1>
+                    {% if result %}
+                        {% for pair in result %}
+                            <div class="game-pair">
+                                <table>
+                                    <thead>
+                                        <tr>
+                                            {% for header in pair[0].keys() %}
+                                                <th>{{ header }}</th>
+                                            {% endfor %}
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        {% for game in pair %}
+                                            <tr>
+                                                {% for value in game.values() %}
+                                                    <td>{{ value }}</td>
+                                                {% endfor %}
+                                            </tr>
+                                        {% endfor %}
+                                    </tbody>
+                                </table>
+                            </div>
+                        {% endfor %}
+                    {% else %}
+                        <p>No data available</p>
+                    {% endif %}
+                </body>
+                </html>
+            """, result=sdql_data)
+        else:
+            scores, odds = get_odds_data(sport_key, selected_date_start)
+            if scores is None or odds is None:
+                return jsonify({'error': 'Error fetching odds data'}), 500
 
-        filtered_scores = []
-        for score in scores:
-            commence_time_str = score['commence_time']
-            commence_date = parser.parse(commence_time_str).astimezone(pytz.utc)
-            if convert_to_eastern(commence_date).date() == selected_date_start.date():
-                filtered_scores.append(score)
+            filtered_scores = []
+            for score in scores:
+                commence_time_str = score['commence_time']
+                commence_date = parser.parse(commence_time_str).astimezone(pytz.utc)
+                commence_date_eastern = convert_to_eastern(commence_date)
 
-        next_game_date = False
-        if not filtered_scores:
-            next_game_date = get_next_game_date_within_7_days(scores, selected_date_start)
+                if commence_date_eastern.date() == selected_date_start.date():
+                    filtered_scores.append(score)
 
-        # ✅ Load pitcher data ONCE if sport is MLB
-        #pitchers_data = {}
-        #if sport_key == 'baseball_mlb':
-            #try:
-                # Only scrape if JSON doesn't exist or isn't from today
-                #json_path = "mlb_starting_pitchers.json"
-                #scrape_today = True
+            next_game_date = False
+            if not filtered_scores:
+                next_game_date = get_next_game_date_within_7_days(scores, selected_date_start)
 
-                #if os.path.exists(json_path):
-                    #with open(json_path, "r") as f:
-                        #data = json.load(f)
-                        #if data and data[0].get("date") == datetime.today().strftime("%Y-%m-%d"):
-                            #scrape_today = False
+            formatted_scores = []
+            for match in filtered_scores:
+                home_team = match.get('home_team', 'N/A')
+                away_team = match.get('away_team', 'N/A')
+                home_score = match['scores'][0]['score'] if match.get('scores') else 'N/A'
+                away_score = match['scores'][1]['score'] if match.get('scores') else 'N/A'
 
-                #if scrape_today:
-                    #get_starting_pitchers()  # Will update JSON
+                match_odds = next((odds_match for odds_match in odds if odds_match['id'] == match['id']), None)
+                odds_data = {'h2h': [], 'spreads': [], 'totals': []}
 
-                # Load pitcher data from JSON
-               # with open(json_path, "r") as f:
-                    #for game in json.load(f):
-                       # key = f"{game['away_team']}@{game['home_team']}"
-                       # pitchers_data[key] = game
-
-            #except Exception as e:
-               # print("Error handling pitcher data:", e)
-
-        formatted_scores = []
-        for match in filtered_scores:
-            home_team = match.get('home_team', 'N/A')
-            away_team = match.get('away_team', 'N/A')
-            home_score = match['scores'][0]['score'] if match.get('scores') else 'N/A'
-            away_score = match['scores'][1]['score'] if match.get('scores') else 'N/A'
-
-            # 🔑 Only MLB: lookup pitcher info
-            #home_pitcher = away_pitcher = home_pitcher_stats = away_pitcher_stats = ''
-            #if sport_key == 'baseball_mlb':
-                #home_abbr = convert_roto_team_names(home_team)
-                #away_abbr = convert_roto_team_names(away_team)
-                #key = f"{away_abbr}@{home_abbr}"
-                #pitcher_info = pitchers_data.get(key, {})
-                #away_pitcher = pitcher_info.get('away_pitcher', '')
-                #away_pitcher_stats = pitcher_info.get('away_pitcher_stats', '')
-                #home_pitcher = pitcher_info.get('home_pitcher', '')
-                #home_pitcher_stats = pitcher_info.get('home_pitcher_stats', '')
-
-            # Odds
-            match_odds = next((o for o in odds if o['id'] == match['id']), None)
-            odds_data = {'h2h': [], 'spreads': [], 'totals': []}
-
-            if match_odds:
-                for bookmaker in match_odds['bookmakers']:
-                    for market in bookmaker['markets']:
-                        market_key = market['key']
-                        for outcome in market['outcomes']:
-                            outcome_text = f"{outcome['name']}"
-                            price = outcome['price']
-                            if price > 0:
-                                outcome_text += f": +{price}"
-                            else:
-                                outcome_text += f": {price}"
-                            if market_key in ['spreads', 'totals'] and 'point' in outcome:
-                                outcome_text += f": {outcome['point']}"
-                            if sport_key == 'soccer_epl' and market_key == 'h2h':
-                                odds_data['h2h'].append(outcome_text)
-                            elif sport_key == 'baseball_mlb':
-                                if market_key in odds_data:
-                                    odds_data[market_key].append(outcome_text)
-                            else:
+                if match_odds:
+                    for bookmaker in match_odds['bookmakers']:
+                        for market in bookmaker['markets']:
+                            market_key = market['key']
+                            for outcome in market['outcomes']:
+                                outcome_text = f"{outcome['name']}"
+                                if market_key in ['spreads', 'totals'] and 'point' in outcome:
+                                    outcome_text += f": {outcome['point']}"
+                                if market_key == 'h2h':
+                                    price = outcome['price']
+                                    if price > 0:
+                                        outcome_text += f": +{price}"
+                                    else:
+                                        outcome_text += f": {price}"
+                                else:
+                                    price = outcome['price']
+                                    if price > 0:
+                                        outcome_text += f" +{price}"
+                                    else:
+                                        outcome_text += f" {price}"
                                 odds_data[market_key].append(outcome_text)
 
-            formatted_scores.append({
-                'homeTeam': home_team,
-                'awayTeam': away_team,
-                'homeScore': away_score,
-                'awayScore': home_score,
-                'odds': odds_data,
-                'game_id': match['id'],
-                #'homePitcher': home_pitcher,
-                #'homePitcherStats': home_pitcher_stats,
-                #'awayPitcher': away_pitcher,
-                #'awayPitcherStats': away_pitcher_stats,
-            })
-            #if sport_key == 'baseball_mlb':
-                #print(f"{away_team} @ {home_team}")
-                #print(f"  Away Pitcher: {away_pitcher} ({away_pitcher_stats})")
-                #print(f"  Home Pitcher: {home_pitcher} ({home_pitcher_stats})")
+                formatted_scores.append({
+                    'homeTeam': home_team,
+                    'awayTeam': away_team,
+                    #cores are purposefully switched investigating why they're backwards
+                    'homeScore': away_score,
+                    'awayScore': home_score,
+                    'odds': odds_data,
+                    'game_id': match['id'],
+                })
 
-        template = baseball_template if sport_key == 'baseball_mlb' else (
-            soccer_template if sport_key == 'soccer_epl' else default_template
-        )
-
-        return render_template_string(
-            template,
-            result=formatted_scores,
-            sport_key=sport_key,
-            current_date=current_date,
-            next_game_date=next_game_date,
-            excluded_sports=EXCLUDED_SPORTS
-        )
+            return render_template_string("""
+                <html>
+                <head>
+                    <title>Game Info</title>
+                    <script 
+                        async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6546677374101814"
+                        crossorigin="anonymous">
+                    </script>
+                    <style>
+                        table {
+                            width: 80%;
+                            border-collapse: collapse;
+                            margin-left: auto;
+                            margin-right: auto;
+                        }
+                        th, td {
+                            padding: 8px;
+                            text-align: left;
+                        }
+                        th {
+                            background-color: #007bff;
+                            color: white;
+                        }
+                        tr:nth-child(even) {
+                            background-color: #d8ebff;
+                        }
+                        tr:nth-child(odd) {
+                            background-color: #FFFFFF;
+                        }
+                        .odds-category {
+                            margin-top: 10px;
+                        }
+                        .odds-category h4 {
+                            margin-bottom: 5px;
+                            color: #007bff;
+                        }
+                        .odds-category ul {
+                            list-style-type: none;
+                            padding: 0;
+                        }
+                        .center {
+                            text-align: center;
+                        }
+                        
+                    </style>
+                </head>
+                <body>
+                    <h1 class="center">Game Information</h1>
+                    {% if result %}
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th>Home Team</th>
+                                    <th>Away Team</th>
+                                    <th>Home Score</th>
+                                    <th>Away Score</th>
+                                    <th>Odds</th>
+                                    {% if sport_key not in excluded_sports %}
+                                        <th>Details</th>
+                                    {% endif %}
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {% for match in result %}
+                                    <tr>
+                                        <td>{{ match.homeTeam }}</td>
+                                        <td>{{ match.awayTeam }}</td>
+                                        <td>{{ match.awayScore }}</td>
+                                        <td>{{ match.homeScore }}</td>
+                                        <td>
+                                            <div class="odds-category">
+                                                <h4>H2H:</h4>
+                                                <ul>
+                                                    {% for odd in match.odds.h2h %}
+                                                        <li>{{ odd }}</li>
+                                                    {% endfor %}
+                                                </ul>
+                                                <h4>Spreads:</h4>
+                                                <ul>
+                                                    {% for odd in match.odds.spreads %}
+                                                        <li>{{ odd }}</li>
+                                                    {% endfor %}
+                                                </ul>
+                                                <h4>Totals:</h4>
+                                                <ul>
+                                                    {% for odd in match.odds.totals %}
+                                                        <li>{{ odd }}</li>
+                                                    {% endfor %}
+                                                </ul>
+                                            </div>
+                                        </td>
+                                        {% if sport_key not in excluded_sports %}
+                                            <td>
+                                                <a href="/game/{{ match.game_id }}?sport_key={{ sport_key }}&date={{ current_date }}" class="view-details">View Details</a>
+                                            </td>
+                                        {% endif %}
+                                    </tr>
+                                {% endfor %}
+                            </tbody>
+                        </table>
+                    {% elif next_game_date %}
+                        <p class="center">No games on this date. The next game is on {{ next_game_date['pretty_date'] }}.</p>
+                        <p class="center"><a href="javascript:void(0);" onclick="goToNextGame('{{ next_game_date['commence_date'] }}')" class="button center">Go to Next Game</a></p>
+                    {% else %}
+                        <p class="center">No games on this date</p>
+                    {% endif %}
+                </body>
+                </html>
+            """, result=formatted_scores, sport_key=sport_key, current_date=current_date, next_game_date=next_game_date, excluded_sports=EXCLUDED_SPORTS)
 
     except requests.exceptions.RequestException as e:
-        print('Request error:', e)
+        print('Request error:', str(e))
         return jsonify({'error': 'Request Error'}), 500
     except Exception as e:
-        print('Error fetching scores:', e)
+        print('Error fetching scores:', str(e))
         return jsonify({'error': 'Internal Server Error'}), 500
 
+
+# @app.route('/trends')
+# @cache.cached(timeout=3600, query_string=True)
+# def show_trends():
+#     sport_key = request.args.get('sport_key')
+#     date = request.args.get('date')
+
+#     if not sport_key or not date:
+#         return jsonify({'error': 'Missing sport_key or date'}), 400
+
+#     try:
+#         selected_date_start = datetime.strptime(date, '%Y-%m-%d').replace(tzinfo=eastern_tz)
+#         scores, odds = get_odds_data(sport_key, selected_date_start)
+#         if scores is None or odds is None:
+#             return jsonify({'error': 'Error fetching odds data'}), 500
+
+#         filtered_scores = []
+#         for score in scores:
+#             commence_time_str = score['commence_time']
+#             commence_date = parser.parse(commence_time_str).astimezone(pytz.utc)
+#             commence_date_eastern = convert_to_eastern(commence_date)
+
+#             if commence_date_eastern.date() == selected_date_start.date():
+#                 filtered_scores.append(score)
+
+#         next_game_date = False
+#         if not filtered_scores:
+#             next_game_date = get_next_game_date_within_7_days(scores, selected_date_start)
+
+#         formatted_scores = []
+#         for match in filtered_scores:
+#             home_team = match.get('home_team', 'N/A')
+#             away_team = match.get('away_team', 'N/A')
+#             home_score = match['scores'][0]['score'] if match.get('scores') else 'N/A'
+#             away_score = match['scores'][1]['score'] if match.get('scores') else 'N/A'
+
+#             match_odds = next((odds_match for odds_match in odds if odds_match['id'] == match['id']), None)
+#             odds_data = {'h2h': [], 'spreads': [], 'totals': []}
+
+#             if match_odds:
+#                 for bookmaker in match_odds['bookmakers']:
+#                     for market in bookmaker['markets']:
+#                         market_key = market['key']
+#                         for outcome in market['outcomes']:
+#                             outcome_text = f"{outcome['name']}"
+#                             if market_key in ['spreads', 'totals'] and 'point' in outcome:
+#                                 outcome_text += f": {outcome['point']}"
+#                             if market_key == 'h2h':
+#                                 price = outcome['price']
+#                                 if price > 0:
+#                                     outcome_text += f": +{price}"
+#                                 else:
+#                                     outcome_text += f": {price}"
+#                             else:
+#                                 price = outcome['price']
+#                                 if price > 0:
+#                                     outcome_text += f" +{price}"
+#                                 else:
+#                                     outcome_text += f" {price}"
+#                             odds_data[market_key].append(outcome_text)
+
+#             formatted_scores.append({
+#                 'homeTeam': home_team,
+#                 'awayTeam': away_team,
+#                 'homeScore': away_score,
+#                 'awayScore': home_score,
+#                 'odds': odds_data,
+#                 'game_id': match['id'],
+#             })
+
+#         # Filter the games to include only those with trends
+#         games_with_trends = [game for game in formatted_scores if check_for_trends(game, selected_date_start, sport_key)['trend_detected']]
+
+#         current_date = request.args.get('date', None)
+#         if not current_date:
+#             current_date = datetime.now(eastern_tz).strftime('%Y-%m-%d')
+
+#         return render_template('trends_list.html', result=games_with_trends, sport_key=sport_key, current_date=current_date)
+#     except Exception as e:
+#         print('Error fetching games with trends:', str(e))
+#         return jsonify({'error': 'Internal Server Error'}), 500
+    
+# def detect_trends(games, sport_key):
+#     def is_winner(points, o_points):
+#         if points is None or o_points is None:
+#             return None
+#         return points > o_points
+
+#     def calculate_line_result(points, line, o_points):
+#         if points is None or line is None or o_points is None:
+#             return None, ''
+#         if points + line > o_points:
+#             return True, 'green-bg'
+#         elif points + line < o_points:
+#             return False, 'red-bg'
+#         else:
+#             return None, ''
+
+#     def other_totals(points, o_points, total):
+#         if points is None or o_points is None or total is None:
+#             return None, ''
+#         if points + o_points > total:
+#             return True, 'green-bg'
+#         elif points + o_points < total:
+#             return False, 'red-bg'
+#         else:
+#             return None, ''
+
+#     if sport_key == 'icehockey_nhl':
+#         points_key = 'goals'
+#     else:
+#         points_key = 'points'
+
+#     # Check for trends in the 'team' column
+#     team_colors = []
+#     for game in games:
+#         result = is_winner(game[points_key], game[f'o:{points_key}'])
+#         if result is not None:
+#             color = 'green-bg' if result else 'red-bg'
+#             team_colors.append(color)
+#     if team_colors.count('green-bg') == 5 or team_colors.count('red-bg') == 5:
+#         return True
+
+#     # Check for trends in the 'points' column
+#     points_colors = []
+#     for game in games:
+#         result = is_winner(game[points_key], game[f'o:{points_key}'])
+#         if result is not None:
+#             color = 'green-bg' if result else 'red-bg'
+#             points_colors.append(color)
+#     if points_colors.count('green-bg') == 5 or points_colors.count('red-bg') == 5:
+#         return True
+
+#     # Skip line trend check for NHL
+#     if sport_key != 'icehockey_nhl':
+#         # Check for trends in the 'line' column
+#         line_colors = []
+#         for game in games:
+#             result, color = calculate_line_result(game[points_key], game['line'], game[f'o:{points_key}'])
+#             if result is not None:
+#                 line_colors.append(color)
+#         if line_colors.count('green-bg') == 5 or line_colors.count('red-bg') == 5:
+#             return True
+
+#     # Check for trends in the 'total' column
+#     total_colors = []
+#     for game in games:
+#         result, color = other_totals(game[points_key], game[f'o:{points_key}'], game['total'])
+#         if result is not None:
+#             total_colors.append(color)
+#     if total_colors.count('green-bg') == 5 or total_colors.count('red-bg') == 5:
+#         return True
+
+#     return False
 
 # Route to fetch and display details for a specific game
 @app.route('/game/<game_id>')
@@ -447,8 +778,7 @@ def game_details(game_id):
         try:
             home_team_last_5 = get_last_5_games(game_details['homeTeam'], selected_date, sport_key) or []
             away_team_last_5 = get_last_5_games(game_details['awayTeam'], selected_date, sport_key) or []
-            #print(f"Home Team Last 5: {home_team_last_5}")
-            #print(f"Away Team Last 5: {away_team_last_5}")
+
             # Fetch the last 5 games between the home team and the away team
             last_5_vs_opponent = get_last_5_games_vs_opponent(
                 team=game_details['homeTeam'],
@@ -547,58 +877,1862 @@ def game_details(game_id):
             except TypeError:
                 return None, None
 
-        #pitchers_data = {}
-        #try:
-            #with open('mlb_starting_pitchers.json', 'r') as f:
-                #for game in json.load(f):
-                    #key = f"{game['away_team']}@{game['home_team']}"
-                    #pitchers_data[key] = game
-        #except Exception as e:
-            #print("Error loading pitcher data:", e)
+        # MLB template rendering
+        mlb_template = render_template_string("""
+           <html>
+                <head>
+                    <!-- Google tag (gtag.js) -->
+                    <script async src="https://www.googletagmanager.com/gtag/js?id=G-578SDWQPSK"></script>
+                    <script>
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
 
-        #away_abbr = convert_roto_team_names(game_details["awayTeam"])
-        #home_abbr = convert_roto_team_names(game_details["homeTeam"])
+                    gtag('config', 'G-578SDWQPSK');
+                    </script> 
+                    <script 
+                        async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6546677374101814"
+                        crossorigin="anonymous">
+                    </script>                         
+                    <title>Game Details</title>
+                    <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}" type="image/x-icon">
+                    <style>
+                        body {
+                            background-color: #f9f9f9; /* Light background */
+                            font-family: 'Poppins', sans-serif;
+                            margin: 0;
+                            padding: 20px;
+                            color: #354050;
+                        }
 
-        #matchup_key = f"{away_abbr}@{home_abbr}"
-        #pitchers = pitchers_data.get(matchup_key, {})
-        #print(f"{away_abbr} @ {home_abbr} → Pitchers: {pitchers}")
+                        header {
+                            background-color: #007bff; /* Primary color */
+                            padding: 10px 20px;
+                            text-align: center;
+                            color: white;
+                        }
 
-        #print(f"Redirecting to game_details for game_id: {game_id}")
-       
+                        nav a {
+                            color: white;
+                            text-decoration: none;
+                            margin: 0 10px;
+                            display: block;                  
+                        }
+
+                        h1 {
+                            margin: 20px 0;
+                        }
+
+                        h2 {
+                            margin-top: 30px;
+                        }
+
+                        table {
+                            width: 90%;
+                            border-collapse: collapse;
+                            margin-bottom: 20px;
+                        }
+
+                        table, th, td {
+                            border: 1.5px solid #ddd;
+                        }
+
+                        th, td {
+                            padding: 4px;
+                            text-align: left;                 
+                        }
+
+                        th {
+                            background-color: #f2f2f2;
+                        }
+
+                        .green-bg {
+                            background-color: #7ebe7e;
+                            color: black;
+                        }
+
+                        .red-bg {
+                            background-color: #e35a69;
+                            color: black;
+                        }
+
+                        .grey-bg {
+                            background-color: #c6c8ca;
+                            color: black;
+                        }
+
+                        #gamesList {
+                            list-style-type: none;
+                            padding: 0;
+                        }
+                
+                        .game-card {
+                            border: 1px solid #ddd;
+                            border-radius: 8px;
+                            padding: 20px;
+                            margin: 10px 0;
+                            background-color: #fff; /* Card background */
+                            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+                        }
+
+                        button, a {
+                            background-color: #007bff; /* Primary color */
+                            color: white;
+                            border: none;
+                            border-radius: 5px;
+                            padding: 10px 15px;
+                            text-decoration: none;
+                            transition: background-color 0.3s;
+                        }
+
+                        button:hover, a:hover {
+                            background-color: #0056b3; /* Darker shade on hover */
+                        }
+
+                        button:disabled {
+                            background-color: #cccccc; /* Light grey background */
+                            color: #666666; /* Dark grey text */
+                            cursor: not-allowed; /* Change cursor to not-allowed */
+                            opacity: 0.6; /* Reduce opacity */
+                        }
+
+                        /* Responsive Styles */
+                        @media only screen and (max-width: 600px) {
+                            body {
+                                font-size: 14px;
+                            }
+                        }
+                        @media only screen and (min-width: 601px) and (max-width: 768px) {
+                            body {
+                                font-size: 16px;
+                            }
+                        }
+                        @media only screen and (min-width: 769px) {
+                            body {
+                                font-size: 18px;
+                            }
+                        }
+                        .info-icon {
+                            cursor: pointer;
+                            color: black; /* Color of the info icon */
+                            margin-left: 1px; /* Space between title and icon */
+                            padding: 1px;
+                            border: 1px solid black; /* Border color */
+                            border-radius: 50%; /* Makes the icon circular */
+                            width: 10px; /* Width of the icon */
+                            height: 10px; /* Height of the icon */
+                            display: inline-flex; /* Allows for centering */
+                            align-items: center; /* Center content vertically */
+                            justify-content: center; /* Center content horizontally */
+                            font-size: 14px; /* Font size of the "i" */
+                        }
+
+                        .info-icon:hover::after {
+                            content: attr(title);
+                            position: absolute;
+                            background: #fff;
+                            border: 1px solid #ccc;
+                            padding: 5px;
+                            border-radius: 4px;
+                            white-space: nowrap;
+                            z-index: 10;
+                            top: 20px; /* Adjust as needed */
+                            left: 0;
+                            box-shadow: 0 0 10px rgba(0,0,0,0.2);
+                        }               
+                        .color-keys {
+                            margin: 20px 0;
+                            padding: 15px;
+                            border: 1px solid #ddd;
+                            background-color: #f9f9f9;
+                        }
+
+                        .color-key h2 {
+                            margin: 0 0 10px;
+                        }
+
+                        .color-key table {
+                            width: 100%;
+                            border-collapse: collapse;
+                        }
+
+                        .color-key th, .color-key td {
+                            padding: 8px;
+                            text-align: left;
+                            border: 1px solid #ccc;
+                        }
+
+                        .color-key th {
+                            background-color: #f2f2f2;
+                        }
+                        .color-key {
+                            display: none; /* Hide by default */
+                            background-color: #f9f9f9; /* Background color for the key */
+                            border: 1px solid #ccc; /* Border for the key */
+                            padding: 5px;
+                            position: absolute; /* Position it relative to the header */
+                            z-index: 10; /* Ensure it appears above other elements */
+                        }
+
+                        th {
+                            position: relative; /* Needed for absolute positioning of the color key */
+                        }
+
+                        th:hover .color-key {
+                            display: block; /* Show the key on hover */
+                        }                      
+                      
+                    </style>
+
+                </head>
+                <body>
+                     <header>
+                        <nav>
+                            <a href="/">Home</a>
+                        </nav>
+                    </header>                         
+                    <h1>Game Details</h1>
+                    <table >
+                        <tr>
+                            <th>Home Team</th>
+                            <td>{{ game.homeTeam }}</td>
+                        </tr>
+                        <tr>
+                            <th>Away Team</th>
+                            <td>{{ game.awayTeam }}</td>
+                        </tr>
+                        <tr>
+                            <th>Home Score</th>
+                            <td>{{ game.homeScore }}</td>
+                        </tr>
+                        <tr>
+                            <th>Away Score</th>
+                            <td>{{ game.awayScore }}</td>
+                        </tr>
+                        <tr>
+                            <th>Odds</th>
+                            <td>{{ game.oddsText.replace('\n', '<br>')|safe }}</td>
+                        </tr>
+                    </table>
+                                              
+                                               
+                    
+                    <h2>Last 5 Games - {{ game.homeTeam }}</h2>
+                    {% if home_team_last_5 %}
+                       <div class="game-card">                       
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Runs <span class="info-icon">i</span>
+                                    
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Runs</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {% for game in home_team_last_5 %}
+                                    {% set is_total_exceeded, total_class = mlb_totals(game.get('runs', 0), game.get('o:runs', 0), game.get('total', 0)) %}
+                                    {% set is_winner = mlb_winner(game.get('runs', 0), game.get('o:runs', 0)) %}
+                                    <tr>
+                                        <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                        <td>{{ game['site'] }}</td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['team'] }}
+                                        </td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['runs'] }}
+                                        </td>
+                                        <td>{{ game['line'] }}</td>
+                                        <td>{{ game['o:team'] }}</td>
+                                        <td>{{ game['o:runs'] }}</td>
+                                        <td class="{{ total_class }}">
+                                            {{ game['total'] }}
+                                        </td>
+                                    </tr>
+                                {% endfor %}
+                            </tbody>
+                        </table>
+                      </div>                        
+                    {% else %}
+                        <p class="center">No data available.</p>
+                    {% endif %}
+                    
+                    <h2>Last 5 Games - {{ game.awayTeam }}</h2>
+                    {% if away_team_last_5 %}
+                      <div class="game-card">                        
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Runs <span class="info-icon">i</span>
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Runs</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {% for game in away_team_last_5 %}
+                                    {% set is_total_exceeded, total_class = mlb_totals(game.get('runs', 0), game.get('o:runs', 0), game.get('total', 0)) %}
+                                    {% set is_winner = mlb_winner(game.get('runs', 0), game.get('o:runs', 0)) %}
+                                    <tr>
+                                        <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                        <td>{{ game['site'] }}</td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['team'] }}
+                                        </td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['runs'] }}
+                                        </td>             
+                                        <td>{{ game['line'] }}</td>
+                                        <td>{{ game['o:team'] }}</td>
+                                        <td>{{ game['o:runs'] }}</td>
+                                        <td class="{{ total_class }}">
+                                            {{ game['total'] }}
+                                        </td>
+                                    </tr>
+                                {% endfor %}
+                            </tbody>
+                        </table>
+                      </div>                        
+                    {% else %}
+                        <p class="center">No data available.</p>
+                    {% endif %}
+                    
+                    <h2>Last 5 Games Between {{ game.homeTeam }} and {{ game.awayTeam }}</h2>
+                    {% if last_5_vs_opponent %}
+                       <div class="game-card">                       
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Runs <span class="info-icon">i</span>
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Runs</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {% for game in last_5_vs_opponent %}
+                                    {% set is_total_exceeded, total_class = mlb_totals(game.get('runs', 0), game.get('o:runs', 0), game.get('total', 0)) %}
+                                    {% set is_winner = mlb_winner(game.get('runs', 0), game.get('o:runs', 0)) %}
+                                    <tr>
+                                        <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                        <td>{{ game['site'] }}</td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['team'] }}
+                                        </td>
+                                        <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                            {{ game['runs'] }}
+                                        </td>
+                                        <td>{{ game['line'] }}</td>
+                                        <td>{{ game['o:team'] }}</td>
+                                        <td>{{ game['o:runs'] }}</td>
+                                        <td class="{{ total_class }}">
+                                            {{ game['total'] }}
+                                        </td>
+                                    </tr>
+                                {% endfor %}
+                            </tbody>
+                        </table>
+                       </div>                        
+                    {% else %}
+                        <p class="center">No data available.</p>
+                    {% endif %}
+                </body>
+                </html>
+            """, game=game_details, home_team_last_5=home_team_last_5, away_team_last_5=away_team_last_5, last_5_vs_opponent=last_5_vs_opponent, mlb_totals=mlb_totals, mlb_winner=mlb_winner)
+        
+        # NHL template rendering
+        nhl_template = render_template_string("""
+            <html>
+            <head>
+                <!-- Google tag (gtag.js) -->
+                    <script async src="https://www.googletagmanager.com/gtag/js?id=G-578SDWQPSK"></script>
+                    <script>
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
+
+                    gtag('config', 'G-578SDWQPSK');
+                    </script>   
+                    <script 
+                        async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6546677374101814"
+                        crossorigin="anonymous">
+                    </script>                           
+                <title>Game Details</title>
+                <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}" type="image/x-icon">
+                <style>
+                        body {
+                            background-color: #f9f9f9; /* Light background */
+                            font-family: 'Poppins', sans-serif;
+                            margin: 0;
+                            padding: 20px;
+                            color: #354050;
+                        }
+
+                        header {
+                            background-color: #007bff; /* Primary color */
+                            padding: 10px 20px;
+                            text-align: center;
+                            color: white;
+                        }
+
+                        nav a {
+                            color: white;
+                            text-decoration: none;
+                            margin: 0 10px;
+                            display: block;                  
+                        }
+
+                        h1 {
+                            margin: 20px 0;
+                        }
+
+                        h2 {
+                            margin-top: 30px;
+                        }
+
+                        table {
+                            width: 90%;
+                            border-collapse: collapse;
+                            margin-bottom: 20px;
+                        }
+
+                        table, th, td {
+                            border: 1.5px solid #ddd;
+                        }
+
+                        th, td {
+                            padding: 4px;
+                            text-align: left;                 
+                        }
+
+                        th {
+                            background-color: #f2f2f2;
+                        }
+
+                        .green-bg {
+                            background-color: #7ebe7e;
+                            color: black;
+                        }
+
+                        .red-bg {
+                            background-color: #e35a69;
+                            color: black;
+                        }
+
+                        .grey-bg {
+                            background-color: #c6c8ca;
+                            color: black;
+                        }
+
+                        #gamesList {
+                            list-style-type: none;
+                            padding: 0;
+                        }
+                
+                        .game-card {
+                            border: 1px solid #ddd;
+                            border-radius: 8px;
+                            padding: 20px;
+                            margin: 10px 0;
+                            background-color: #fff; /* Card background */
+                            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+                        }
+                                              
+                        button:disabled {
+                            background-color: #cccccc; /* Light grey background */
+                            color: #666666; /* Dark grey text */
+                            cursor: not-allowed; /* Change cursor to not-allowed */
+                            opacity: 0.6; /* Reduce opacity */
+                        }
+
+                        button, a {
+                            background-color: #007bff; /* Primary color */
+                            color: white;
+                            border: none;
+                            border-radius: 5px;
+                            padding: 10px 15px;
+                            text-decoration: none;
+                            transition: background-color 0.3s;
+                        }
+
+                        button:hover, a:hover {
+                            background-color: #0056b3; /* Darker shade on hover */
+                        }
+
+                        /* Responsive Styles */
+                        @media only screen and (max-width: 600px) {
+                            body {
+                                font-size: 14px;
+                            }
+                        }
+                        @media only screen and (min-width: 601px) and (max-width: 768px) {
+                            body {
+                                font-size: 16px;
+                            }
+                        }
+                        @media only screen and (min-width: 769px) {
+                            body {
+                                font-size: 18px;
+                            }
+                        }
+                        .info-icon {
+                            cursor: pointer;
+                            color: black; /* Color of the info icon */
+                            margin-left: 1px; /* Space between title and icon */
+                            padding: 1px;
+                            border: 1px solid black; /* Border color */
+                            border-radius: 50%; /* Makes the icon circular */
+                            width: 10px; /* Width of the icon */
+                            height: 10px; /* Height of the icon */
+                            display: inline-flex; /* Allows for centering */
+                            align-items: center; /* Center content vertically */
+                            justify-content: center; /* Center content horizontally */
+                            font-size: 14px; /* Font size of the "i" */
+                        }
+
+                        .info-icon:hover::after {
+                            content: attr(title);
+                            position: absolute;
+                            background: #fff;
+                            border: 1px solid #ccc;
+                            padding: 5px;
+                            border-radius: 4px;
+                            white-space: nowrap;
+                            z-index: 10;
+                            top: 20px; /* Adjust as needed */
+                            left: 0;
+                            box-shadow: 0 0 10px rgba(0,0,0,0.2);
+                        }               
+                        .color-keys {
+                            margin: 20px 0;
+                            padding: 15px;
+                            border: 1px solid #ddd;
+                            background-color: #f9f9f9;
+                        }
+
+                        .color-key h2 {
+                            margin: 0 0 10px;
+                        }
+
+                        .color-key table {
+                            width: 100%;
+                            border-collapse: collapse;
+                        }
+
+                        .color-key th, .color-key td {
+                            padding: 8px;
+                            text-align: left;
+                            border: 1px solid #ccc;
+                        }
+
+                        .color-key th {
+                            background-color: #f2f2f2;
+                        }
+                        .color-key {
+                            display: none; /* Hide by default */
+                            background-color: #f9f9f9; /* Background color for the key */
+                            border: 1px solid #ccc; /* Border for the key */
+                            padding: 5px;
+                            position: absolute; /* Position it relative to the header */
+                            z-index: 10; /* Ensure it appears above other elements */
+                        }
+
+                        th {
+                            position: relative; /* Needed for absolute positioning of the color key */
+                        }
+
+                        th:hover .color-key {
+                            display: block; /* Show the key on hover */
+                        }                      
+                      
+                    </style>
+            </head>
+            <body>
+                <header>
+                        <nav>
+                            <a href="/">Home</a>
+                        </nav>
+                    </header>                               
+                <h1>Game Details</h1>
+                <table>
+                    <tr>
+                        <th>Home Team</th>
+                        <td>{{ game.homeTeam }}</td>
+                    </tr>
+                    <tr>
+                        <th>Away Team</th>
+                        <td>{{ game.awayTeam }}</td>
+                    </tr>
+                    <tr>
+                        <th>Home Score</th>
+                        <td>{{ game.homeScore }}</td>
+                    </tr>
+                    <tr>
+                        <th>Away Score</th>
+                        <td>{{ game.awayScore }}</td>
+                    </tr>
+                    <tr>
+                        <th>Odds</th>
+                        <td>{{ game.oddsText.replace('\n', '<br>')|safe }}</td>
+                    </tr>
+                </table>
+                <div><p>Hover over column titles for color meanings</p></div>
+                <h2>Last 5 Games - {{ game.homeTeam }}</h2>
+                {% if home_team_last_5 %}
+                    <table>
+                        <thead>
+                            <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Goals <span class="info-icon">i</span>
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Goals</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in home_team_last_5 %}
+                                {% set is_total_exceeded, total_class = nhl_totals(game.get('goals', 0), game.get('o:goals', 0), game.get('total', 0)) %}
+                                {% set is_winner = nhl_winner(game.get('goals', 0), game.get('o:goals', 0)) %}
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['team'] }}
+                                    </td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['goals'] }}
+                                    </td>
+                                    <td>{{ game['line'] }}</td>
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:goals'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+
+                <h2>Last 5 Games - {{ game.awayTeam }}</h2>
+                {% if away_team_last_5 %}
+                    <table>
+                        <thead>
+                            <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Goals <span class="info-icon">i</span>
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Goals</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in away_team_last_5 %}
+                                {% set is_total_exceeded, total_class = nhl_totals(game.get('goals', 0), game.get('o:goals', 0), game.get('total', 0)) %}
+                                {% set is_winner = nhl_winner(game.get('goals', 0), game.get('o:goals', 0)) %}
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['team'] }}
+                                    </td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['goals'] }}
+                                    </td>
+                                    <td>{{ game['line'] }}</td>
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:goals'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+
+                <h2>Last 5 Games Between {{ game.homeTeam }} and {{ game.awayTeam }}</h2>
+                {% if last_5_vs_opponent %}
+                    <table>
+                        <thead>
+                            <tr>
+                                    <th>Date</th>
+                                    <th>Site</th>
+                                    <th>Team <span class="info-icon">i</span>
+                                       <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>       
+                                    </th>
+                                    <th>Goals <span class="info-icon">i</span>
+                                        <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>      
+                                    </th>
+                                    <th>Line</th>
+                                    <th>Opponent</th>
+                                    <th>Opponent Goals</th>
+                                    <th>Total <span class="info-icon">i</span>
+                                         <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>     
+                                    </th>
+                                </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in last_5_vs_opponent %}
+                                {% set is_total_exceeded, total_class = nhl_totals(game.get('goals', 0), game.get('o:goals', 0), game.get('total', 0)) %}
+                                {% set is_winner = nhl_winner(game.get('goals', 0), game.get('o:goals', 0)) %}
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">{{ game['team'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">{{ game['goals'] }}</td>             
+                                    <td>{{ game['line'] }}</td>
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:goals'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+            </body>
+            </html>
+        """, game=game_details, home_team_last_5=home_team_last_5, away_team_last_5=away_team_last_5, last_5_vs_opponent=last_5_vs_opponent, nhl_totals=nhl_totals, nhl_winner=nhl_winner)
+
+        # Other sports template rendering
+        others_template = render_template_string("""
+            <html>
+            <head>
+                <!-- Google tag (gtag.js) -->
+                    <script async src="https://www.googletagmanager.com/gtag/js?id=G-578SDWQPSK"></script>
+                    <script>
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
+
+                    gtag('config', 'G-578SDWQPSK');
+                    </script> 
+                    <script 
+                        async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6546677374101814"
+                        crossorigin="anonymous">
+                    </script>                                
+                <title>Game Details</title>
+                <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}" type="image/x-icon">
+                <style>
+                        body {
+                            background-color: #f9f9f9; /* Light background */
+                            font-family: 'Poppins', sans-serif;
+                            margin: 0;
+                            padding: 20px;
+                            color: #354050;
+                        }
+
+                        header {
+                            background-color: #007bff; /* Primary color */
+                            padding: 10px 20px;
+                            text-align: center;
+                            color: white;
+                            border-radius: 5px;                     
+                        }
+
+                        nav a {
+                            color: white;
+                            text-decoration: none;
+                            margin: 0 10px;
+                            display: block;                     
+                        }
+
+                        h1 {
+                            margin: 20px 0;
+                        }
+
+                        h2 {
+                            margin-top: 30px;
+                        }
+
+                        table {
+                            width: 90%;
+                            border-collapse: collapse;
+                            margin-bottom: 20px;
+                        }
+
+                        table, th, td {
+                            border: 1.5px solid #ddd;
+                        }
+
+                        th, td {
+                            padding: 4px;
+                            text-align: left;                 
+                        }
+
+                        th {
+                            background-color: #f2f2f2;
+                        }
+
+                        .green-bg {
+                            background-color: #7ebe7e;
+                            color: black;
+                        }
+
+                        .red-bg {
+                            background-color: #e35a69;
+                            color: black;
+                        }
+
+                        .grey-bg {
+                            background-color: #c6c8ca;
+                            color: black;
+                        }
+
+                        #gamesList {
+                            list-style-type: none;
+                            padding: 0;
+                        }
+                
+                        .game-card {
+                            border: 1px solid #ddd;
+                            border-radius: 8px;
+                            padding: 20px;
+                            margin: 10px 0;
+                            background-color: #fff; /* Card background */
+                            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+                        }
+                        
+                        button:disabled {
+                            background-color: #cccccc; /* Light grey background */
+                            color: #666666; /* Dark grey text */
+                            cursor: not-allowed; /* Change cursor to not-allowed */
+                            opacity: 0.6; /* Reduce opacity */
+                        }
+
+                        button, a {
+                            background-color: #007bff; /* Primary color */
+                            color: white;
+                            border: none;
+                            border-radius: 5px;
+                            padding: 10px 15px;
+                            text-decoration: none;
+                            transition: background-color 0.3s;
+                        }
+
+                        button:hover, a:hover {
+                            background-color: #0056b3; /* Darker shade on hover */
+                        }
+                        .info-icon {
+                            cursor: pointer;
+                            color: black; /* Color of the info icon */
+                            margin-left: 1px; /* Space between title and icon */
+                            padding: 1px;
+                            border: 1px solid black; /* Border color */
+                            border-radius: 50%; /* Makes the icon circular */
+                            width: 10px; /* Width of the icon */
+                            height: 10px; /* Height of the icon */
+                            display: inline-flex; /* Allows for centering */
+                            align-items: center; /* Center content vertically */
+                            justify-content: center; /* Center content horizontally */
+                            font-size: 14px; /* Font size of the "i" */
+                        }
+
+                        .info-icon:hover::after {
+                            content: attr(title);
+                            position: absolute;
+                            background: #fff;
+                            border: 1px solid #ccc;
+                            padding: 5px;
+                            border-radius: 4px;
+                            white-space: nowrap;
+                            z-index: 10;
+                            top: 20px; /* Adjust as needed */
+                            left: 0;
+                            box-shadow: 0 0 10px rgba(0,0,0,0.2);
+                        }                         
+                                                 
+                        .color-keys {
+                            margin: 20px 0;
+                            padding: 15px;
+                            border: 1px solid #ddd;
+                            background-color: #f9f9f9;
+                        }
+
+                        .color-key h2 {
+                            margin: 0 0 10px;
+                        }
+
+                        .color-key table {
+                            width: 100%;
+                            border-collapse: collapse;
+                        }
+
+                        .color-key th, .color-key td {
+                            padding: 8px;
+                            text-align: left;
+                            border: 1px solid #ccc;
+                        }
+
+                        .color-key th {
+                            background-color: #f2f2f2;
+                        }
+                        .color-key {
+                            display: none; /* Hide by default */
+                            background-color: #f9f9f9; /* Background color for the key */
+                            border: 1px solid #ccc; /* Border for the key */
+                            padding: 5px;
+                            position: absolute; /* Position it relative to the header */
+                            z-index: 10; /* Ensure it appears above other elements */
+                        }
+
+                        th {
+                            position: relative; /* Needed for absolute positioning of the color key */
+                        }
+
+                        th:hover .color-key {
+                            display: block; /* Show the key on hover */
+                        }                         
+
+                        /* Responsive Styles */
+                        @media only screen and (max-width: 600px) {
+                            body {
+                                font-size: 14px;
+                            }
+                        }
+                        @media only screen and (min-width: 601px) and (max-width: 768px) {
+                            body {
+                                font-size: 16px;
+                            }
+                        }
+                        @media only screen and (min-width: 769px) {
+                            body {
+                                font-size: 18px;
+                            }
+                        }
+                    </style>
+            </head>
+            <body>
+                <header>
+                  <nav>
+                    <a href="/">Home</a>
+                    </nav>
+                </header> 
+                <h1>Game Details</h1>
+                <table>
+                    <tr>
+                        <th>Home Team</th>
+                        <td>{{ game.homeTeam }}</td>
+                    </tr>
+                    <tr>
+                        <th>Away Team</th>
+                        <td>{{ game.awayTeam }}</td>
+                    </tr>
+                    <tr>
+                        <th>Home Score</th>
+                        <td>{{ game.homeScore }}</td>
+                    </tr>
+                    <tr>
+                        <th>Away Score</th>
+                        <td>{{ game.awayScore }}</td>
+                    </tr>
+                    <tr>
+                        <th>Odds</th>
+                        <td>{{ game.oddsText.replace('\n', '<br>')|safe }}</td>
+                    </tr>
+                </table>
+
+                <h2>Last 5 Games - {{ game.homeTeam }}</h2>
+                {% if home_team_last_5 %}
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>Date</th>
+                                <th>Site</th>
+                                <th>Team <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Points <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>                 
+                                <th>Spread <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Spread was covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Spread was not covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Spread was a push</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Opponent</th>
+                                <th>Opponent Points</th>
+                                <th>Total <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in home_team_last_5 %}
+                                {% set is_total_exceeded, total_class = other_totals(game.get('points', 0), game.get('o:points', 0), game.get('total', 0)) %}
+                                {% set is_winner = other_winner(game.get('points', 0), game.get('o:points', 0)) %}
+                                {% set line_win, line_class = calculate_line_result(game.get('points'), game.get('line'), game.get('o:points')) %}                 
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['team'] }}
+                                    </td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['points'] }}
+                                    </td>
+                                    <td class="{{ line_class }}">
+                                        {{ game['line'] }}
+                                    </td>             
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:points'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+
+                <h2>Last 5 Games - {{ game.awayTeam }}</h2>
+                {% if away_team_last_5 %}
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>Date</th>
+                                <th>Site</th>
+                                <th>Team <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Points <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>                 
+                                <th>Spread <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Spread was covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Spread was not covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Spread was a push</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Opponent</th>
+                                <th>Opponent Points</th>
+                                <th>Total <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in away_team_last_5 %}
+                                {% set is_total_exceeded, total_class = other_totals(game.get('points', 0), game.get('o:points', 0), game.get('total', 0)) %}
+                                {% set is_winner = other_winner(game.get('points', 0), game.get('o:points', 0)) %}
+                                {% set line_win, line_class = calculate_line_result(game.get('points'), game.get('line'), game.get('o:points')) %}
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['team'] }}
+                                    </td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">
+                                        {{ game['points'] }}
+                                    </td>             
+                                    <td class="{{ line_class }}">
+                                        {{ game['line'] }}
+                                    </td>
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:points'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+
+                <h2>Last 5 Games Between {{ game.homeTeam }} and {{ game.awayTeam }}</h2>
+                {% if last_5_vs_opponent %}
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>Date</th>
+                                <th>Site</th>
+                                <th>Team <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>can I
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Points <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Team won</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Team lost</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>                 
+                                <th>Spread <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Spread was covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Spread was not covered</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Spread was a push</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                                <th>Opponent</th>
+                                <th>Opponent Points</th>
+                                <th>Total <span class="info-icon">i</span>
+                                    <span class="color-key">
+                                              <div class="color-keys">
+                                                    <table>
+                                                        <thead>
+                                                            <tr>
+                                                                <th>Color</th>
+                                                                <th>Meaning</th>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr>
+                                                                <td style="background-color: green;">&nbsp;</td>
+                                                                <td>Total went over</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: red;">&nbsp;</td>
+                                                                <td>Total went under</td>
+                                                            </tr>
+                                                            <tr>
+                                                                <td style="background-color: grey;">&nbsp;</td>
+                                                                <td>Push (tie with total)</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </div> 
+                                        </span>             
+                                </th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {% for game in last_5_vs_opponent %}
+                                {% set is_total_exceeded, total_class = other_totals(game.get('points', 0), game.get('o:points', 0), game.get('total', 0)) %}
+                                {% set is_winner = other_winner(game.get('points', 0), game.get('o:points', 0)) %}
+                                {% set line_win, line_class = calculate_line_result(game.get('points'), game.get('line'), game.get('o:points')) %}
+                                <tr>
+                                    <td>{{ game['date'] }}</td>  <!-- Date will be formatted in the template -->
+                                    <td>{{ game['site'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">{{ game['team'] }}</td>
+                                    <td class="{{ 'green-bg' if is_winner else 'red-bg' if is_winner == False else '' }}">{{ game['points'] }}</td>             
+                                    <td class="{{ line_class }}">
+                                        {{ game['line'] }}
+                                    </td>
+                                    <td>{{ game['o:team'] }}</td>
+                                    <td>{{ game['o:points'] }}</td>
+                                    <td class="{{ total_class }}">
+                                        {{ game['total'] }}
+                                    </td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                {% else %}
+                    <p class="center">No data available.</p>
+                {% endif %}
+            </body>
+            </html>
+        """, game=game_details, home_team_last_5=home_team_last_5, away_team_last_5=away_team_last_5, last_5_vs_opponent=last_5_vs_opponent, other_totals=other_totals, other_winner=other_winner, calculate_line_result=calculate_line_result)    
+
         if sport_key == 'baseball_mlb':
-            return render_template_string(mlb_template,
-                                        game=game_details,
-                                        home_team_last_5=home_team_last_5,
-                                        away_team_last_5=away_team_last_5,
-                                        last_5_vs_opponent=last_5_vs_opponent,
-                                        mlb_totals=mlb_totals,
-                                        mlb_winner=mlb_winner,
-                                        calculate_line_result=calculate_line_result)
-                                        #home_pitcher=pitchers.get('home_pitcher'),
-                                        #home_stats=pitchers.get('home_pitcher_stats'),
-                                        #away_pitcher=pitchers.get('away_pitcher'),
-                                        #away_stats=pitchers.get('away_pitcher_stats'))
+            return mlb_template 
         elif sport_key == 'icehockey_nhl':
-            return render_template_string(nhl_template,
-                                        game=game_details,
-                                        home_team_last_5=home_team_last_5,
-                                        away_team_last_5=away_team_last_5,
-                                        last_5_vs_opponent=last_5_vs_opponent,
-                                        nhl_totals=nhl_totals,
-                                        nhl_winner=nhl_winner,
-                                        calculate_line_result=calculate_line_result)
+            return nhl_template         
         elif sport_key in ['americanfootball_nfl', 'americanfootball_ncaaf', 'basketball_nba', 'basketball_ncaab']:
-            return render_template_string(others_template,
-                                        game=game_details,
-                                        home_team_last_5=home_team_last_5,
-                                        away_team_last_5=away_team_last_5,
-                                        last_5_vs_opponent=last_5_vs_opponent,
-                                        other_totals=other_totals,
-                                        other_winner=other_winner,
-                                        calculate_line_result=calculate_line_result)
+            return others_template
         else:
             raise ValueError(f"Unsupported league: {sport_key}")
-
 
     except Exception as e:
         print('Error fetching game details:', str(e))
@@ -619,4 +2753,3 @@ if __name__ == '__main__':
 
     # Ensure the Celery worker is terminated when the Flask app exits
     celery_process.terminate()
-
